refactor: Przeniesienie przycisków i poszerzenie formularza logowania
Przenosi przyciski "Napiwek", "Zmiana języka" i "Info" z panelu bocznego (Sidebar) do nowego, wyśrodkowanego panelu w dolnym pasku (BottomBar), umieszczonego pod paskiem postępu.

Dodaje nowy przycisk "Robert" z ikoną robota do tego samego panelu w dolnym pasku.

Poszerza formularz logowania, zmniejszając jego poziome dopełnienie (padding), aby lepiej wykorzystać przestrzeń na ekranie, zgodnie z prośbą użytkownika.
